Add files via upload
this is a game that two players can play around by their moves one after the another. there are three limited moves so that the winner is going to be decided after succesfull arrangement of their pans by vertical, horizontal or diagonal. and there is chance of getting tie is also possible.